refactor(daemon): split module_access transfer into included parts
